+++ R/corrr.R
@@ -1,0 +1,26 @@
+#' @importFrom stats dist
+#' @importFrom purrr map_df
+#' @importFrom purrr map_dfr
+#' @importFrom purrr map_dfc
+#' @importFrom purrr map_dbl
+#' @importFrom purrr map_int
+#' @importFrom purrr map_lgl
+#' @importFrom purrr map_chr
+#' @importFrom purrr map
+#' @importFrom purrr map2
+#' @importFrom purrr imap_dfr
+#' @importFrom purrr set_names
+#' @importFrom purrr transpose
+#' @importFrom dplyr select
+#' @importFrom dplyr mutate_all
+#' @importFrom dplyr filter
+#' @importFrom dplyr summarise
+#' @importFrom dplyr collect
+#' @importFrom tibble tibble
+#' @importFrom tibble as_tibble
+#' @import ggplot2
+#' @import rlang
+#' @keywords internal
+"_PACKAGE"
+NULL
+utils::globalVariables(c("id", "r", "n"))