+++ inst/doc/databases.R
@@ -1,0 +1,21 @@
+## ----setup, include = FALSE----------------------------------------------
+library(dplyr)
+library(dbplyr)
+library(corrr)
+knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
+
+## ------------------------------------------------------------------------
+con <- DBI::dbConnect(RSQLite::SQLite(), path = ":dbname:")
+db_mtcars <- copy_to(con, mtcars)
+
+## ------------------------------------------------------------------------
+library(dplyr)
+library(corrr)
+db_mtcars %>%
+  correlate(quiet = TRUE)
+
+## ------------------------------------------------------------------------
+db_mtcars %>%
+  correlate(quiet = TRUE) %>%
+  rplot()
+